automatically backed up by learn

+++ index.js
@@ -39,8 +39,8 @@ function makeInt(n){
 }
 
 function preserveDecimal(n){
-  var n = parseFloat('2.222', 10)
-  return parseFloat(n)
+  var n = parseInt('2.222', 10)
+  return parseInt(n)
 }
 
 function preserveDecimal(n){
